requirements.txt
django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-debug-toolbar==4.4.6
django-redis==5.4.0
requests==2.32.3
celery==5.4.0
django-celery-beat==2.6.0
drf-spectacular==0.27.2
stripe==10.7.0
flake8==7.1.1
